backend/requirements.txt
Django>=4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-database-url==3.1.2
Pillow
python-dotenv==1.0.1
whitenoise==6.12.0
gunicorn==23.0.0
psycopg2-binary==2.9.10; sys_platform != 'win32'
